requirements.txt
torch~=1.12.0
torchvision~=0.13.0
numpy~=1.23.1
fastapi~=0.79.0
matplotlib~=3.4.2
opencv-python~=4.6.0.66
Pillow~=9.2.0
requests~=2.28.1
PyYAML~=6.0
layoutparser[ocr]
pytesseract